Mistral AI: test package restructuring. Fix failing tests. Add AOT tests
